+++ src/driver/mongodb/MongoQueryRunner.ts
@@ -406,7 +406,7 @@ export class MongoQueryRunner implements QueryRunner {
     /**
      * Insert a new row with given values into the given table.
      * Returns value of inserted object id.
-     */
+
     async insert(collectionName: string, keyValues: ObjectLiteral): Promise<any> { // todo: fix any
         const results = await this.databaseConnection
             .collection(collectionName)
@@ -416,20 +416,20 @@ export class MongoQueryRunner implements QueryRunner {
             result: results,
             generatedMap: generatedMap
         };
-    }
+    }*/
 
     /**
      * Updates rows that match given conditions in the given table.
-     */
+
     async update(collectionName: string, valuesMap: ObjectLiteral, conditions: ObjectLiteral): Promise<any> { // todo: fix any
         await this.databaseConnection
             .collection(collectionName)
             .updateOne(conditions, valuesMap);
-    }
+    }*/
 
     /**
      * Deletes from the given table by a given conditions.
-     */
+
     async delete(collectionName: string, conditions: ObjectLiteral|ObjectLiteral[]|string, maybeParameters?: any[]): Promise<any> { // todo: fix any
         if (typeof conditions === "string")
             throw new Error(`String condition is not supported by MongoDB driver.`);
@@ -437,7 +437,7 @@ export class MongoQueryRunner implements QueryRunner {
         await this.databaseConnection
             .collection(collectionName)
             .deleteOne(conditions);
-    }
+    }*/
 
     /**
      * Inserts rows into the closure table.