+++ test/functional/persistence/cascades/cascades-example1/cascades-example1.ts
@@ -27,7 +27,7 @@ describe("persistence > cascades > example 1", () => {
         await connection.manager.save(user);
 
         const loadedUser = await connection.manager
-            .createQueryBuilder(User,"user")
+            .createQueryBuilder(User, "user")
             .leftJoinAndSelect("user.profile", "profile")
             .leftJoinAndSelect("profile.photo", "profilePhoto")
             .leftJoinAndSelect("profile.user", "profileUser")