+++ test/github-issues/836/issue-836.ts
@@ -36,7 +36,7 @@ describe("github issues > #836 .save won't update entity when it contains OneToO
         await connection.manager.save(credential);
 
         // check if credentials and user are saved properly
-        const loadedCredentials = await connection.manager.findOne(UserCredential,2, { relations: ["user"] });
+        const loadedCredentials = await connection.manager.findOne(UserCredential, 2, { relations: ["user"] });
         loadedCredentials!.should.be.eql({
             user: {
                 id: 2,